requirements.txt
boto3
numpy
tensorflow
mysql-connector-python
librosa
Keras